Struct vectis_proxy::contract::VectisProxy

pub struct VectisProxy<'a> {
    pub created_at: Item<'a, u64>,
    pub deployer: Deployer<'a>,
    pub controller: ControllerState<'a>,
    pub vid: Item<'a, String>,
    pub data: ProxyData<'a>,
    pub relayers: Map<'a, Addr, ()>,
    pub exec_plugins: Map<'a, &'a str, PluginInfo>,
    pub pre_tx_plugins: Map<'a, &'a str, PluginInfo>,
    pub post_tx_hooks_plugins: Map<'a, &'a str, PluginInfo>,
    pub installed_plugins: Map<'a, &'a u64, ()>,
    pub previous_controllers: Map<'a, &'a [u8], ()>,
}

Fields
created_at: Item<'a, u64>deployer: Deployer<'a>

The deployer of the Vectis infra

controller: ControllerState<'a>

The Controller of this wallet

vid: Item<'a, String>

Vectis ID of the wallet tied to username hashed

data: ProxyData<'a>

Arbitrary KV store for data written to on instantiation and updatable by users

relayers: Map<'a, Addr, ()>

This records the current relayers that has feegrants

exec_plugins: Map<'a, &'a str, PluginInfo>

Plugins that can exec through this contract unconditionally. The plugin should be from a trusted source managing all permissions This is similar to granting authz to the plugin Example of this: automation, allowance, session keys (assuming each plugin can only have 1 permission) We separate these so that we can only loop through the required ones at pre / post hooks

pre_tx_plugins: Map<'a, &'a str, PluginInfo>post_tx_hooks_plugins: Map<'a, &'a str, PluginInfo>installed_plugins: Map<'a, &'a u64, ()>

allow to check if a plugin has been installed quickly This is the plugin id from vectis registry

previous_controllers: Map<'a, &'a [u8], ()>

The history of the Controllers

Implementations

impl VectisProxy<'_>

pub const fn new() -> Self