Module vectis_proxy::state

Structs

Controller

GuardiansUpdateRequest

Enums

RelayTxError

Relay transaction related errors

Constants

ADDR_PREFIX

Chain address prefix

CODE_ID

CONTROLLER

DAO

The DAO address where the fees for new wallet goes to, also the admin of the contract. On a remote chain, this is the remote-tunnel contract

FROZEN

GUARDIANS

INSTALL_FEE

LABEL

MULTISIG_ADDRESS

PENDING_GUARDIAN_ROTATION

PLUGINS

Plugins

PROXY_MULTISIG_CODE_ID

The latest default multisig code id stored onchain for the proxy

RELAYERS

Type Definitions

Nonce

Controller nonce