Module vectis_proxy::contract

Functions

execute

execute_add_relayer

Add relayer to the relayers set

execute_execute

execute_inst_plugin

Executes instantiation of plugin

execute_plugin_msgs

Call by plugins

execute_relay

Executes relayed message fro a relayer

execute_remove_relayer

Remove relayer from the relayers set

execute_request_update_guardians

execute_revert_freeze_status

Change current freezing status to its inverse Must be from a guardian or a guardian multisig contract

execute_rotate_controller_key

Complete controller key rotation by changing it’s address Must be from a guardian or a guardian multisig contract

execute_update_guardians

execute_update_label

Update label by controller

execute_update_plugin

Add without instantiate, migrate or remove plugin

instantiate

migrate

query

query_can_execute_relay

Returns if sender is a relayer

query_guardian_update_request

query_info

Returns wallet info

query_plugins

reply