Module vectis_proxy::helpers

Functions

add_plugin_to_state

addresses_to_voters

authorize_controller_or_guardians

Is used to authorize controller or guardians

authorize_guardian_or_multisig

Is used to authorize guardian or multisig contract

create_rotate_guardian_factory_msg

ensure_is_contract_self

Ensures sender is relayed from current contract

ensure_is_controller

Ensures sender is the wallet controller

ensure_is_guardian

Ensures sender is guardian

ensure_is_relayer

Ensures sender is relayer

ensure_is_relayer_or_controller

ensure this is either a direct message from the controller or ensure this is relayed by a relayer from this proxy

is_frozen

is_guardian

Checks if the sender is a guardian

is_relayer

Checks if the sender is a relayer

load_addresses

Load addresses from store

load_canonical_addresses

Load canonical addresses from store