Module vectis_proxy::state

Structs

Controller

GuardiansUpdateRequest

Enums

RelayTxError

Relay transaction related errors

Constants

ADDR_PREFIX

Chain address prefix

CODE_ID

CONTROLLER

Vecits Proxy Controller

CREATED_AT

DEPLOYER

In beta-V1 this is a multisig In progressive decentralisation this will be a DAAO

EXEC_PLUGINS

Plugins with execution rights

FROZEN

GUARDIANS

INSTALL_FEE

LABEL

MULTISIG_ADDRESS

MULTISIG_PLUGIN

Plugins that is multisig

PENDING_GUARDIAN_ROTATION

PENDING_MULTISIG

PENDING_PLUGIN

PRE_TX_PLUGINS

Plugins that do tx Pre-checks

PROXY_MULTISIG_CODE_ID

The latest default multisig code id stored onchain for the proxy

QUERY_PLUGINS

Vectis Proxy state for other contract to query it

RELAYERS

Type Definitions

Nonce

Controller nonce